refactor: hotplace 모듈화 기능: 렌더를 따로 모듈화 진행 #3
